+++ changelog.md
@@ -5,8 +5,9 @@
 ### Patch
 
 - Updated dependencies.
-- Add a test for `GraphQL.operate()` with both `reloadOnLoad` and `resetOnLoad` options true.
+- Test `GraphQL.operate()` with both `reloadOnLoad` and `resetOnLoad` options true.
 - Use string `FormData` field names, as some `FormData` polyfills don't coerce numbers like native implementations do.
+- Test files in variables result in appropriate fetch options for a valid [GraphQL multipart request](https://github.com/jaydenseric/graphql-multipart-request-spec).
 
 ## 8.2.0